ant: refactor fetcher
The fetcher is now an implementation that builds on top of the ant.Client
interface, which is implemented by the net/http package.

It allows the fetcher to make higher-level decisions such as when to cache
and when to adjust the charset encoding.

It also allows the fetcher to be reused by a future antcdp package which will
implemented an `ant.Client` of its own.