antcdp: refactor, optimize
Changed antcdp.Client to use a *transport which in turn uses several data-structures
to ensure that we minimize roundtrips, target creation and connecting to the debugger
address.

Adjusted *tx to add cookies when the request contains any and read cookies
from chrome, this allows us to use a cookie jar.